requirements.txt
torch
torchvision
timm
albumentations
opencv-python
numpy
pandas
openpyxl
scikit-learn
segmentation-models-pytorch
streamlit
Pillow
tqdm
fastapi
uvicorn[standard]
python-multipart
